requirements.txt
django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
pandas==1.5.2
